+++ CHANGELOG.md
@@ -5,9 +5,10 @@
 * New builtin functions: `\lastValue`, `\shortestValue`, `\unmul`, `\unsnoc0`.
 * `\orApply`, `\iterate`, `\nTimes`, `\while`, `\lengthWhile`, `\fixedPoint` now support functions with arity `m -> n` where `m >= n`. When `m > n`, the top `m - n` values on the stack are "quoted" as a new function that pushes these values to the stack, and this new function is composed with the original function before applying it.
 * A new mode `last` is added. It outputs the last possible result. You can switch to this mode by the `-t` flag in the command line or the `\Mode last` command in the REPL.
+* `\lengthIs` is now automatically vectorized.
 * Now input can have leading and trailing spaces. The interpreter will trim them when parsing.
-* Now Nekomata can be built with GHC 9.10.1, thanks to an update of the `haskeline` package.
 * Fixed a bug in the REPL. Now the REPL can correctly handle strings with whitespaces.
+* Now Nekomata can be built with GHC 9.10.2 and 9.12.2, thanks to an update of the `haskeline` package.
 
 ### Breaking changes